@@ -10,15 +10,15 @@
     "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.1.0",
+    "@sveltejs/vite-plugin-svelte": "4.0.0-next.7",
     "@tsconfig/svelte": "^5.0.0",
     "autoprefixer": "^10.4.16",
     "postcss": "^8.4.33",
-    "svelte": "^4.2.17",
+    "svelte": "5.0.0-next.262",
     "svelte-check": "^3.7.1",
     "tailwindcss": "^3.4.1",
     "tslib": "^2.6.0",
-    "typescript": "^5.4.5",
+    "typescript": "^5.6.2",
     "vite": "^5.2.11"
   }
 }
